move files back to widget directory
