+++ packages/itwin/map-layers/src/ui/widget/MapLayerSettingsMenu.tsx
@@ -6,10 +6,10 @@ import "../MapLayerManager.scss";
 import React from "react";
 import { SvgMoreVertical } from "@itwin/itwinui-icons-react";
 import { DropdownMenu, IconButton, MenuItem, Slider } from "@itwin/itwinui-react";
-import { MapLayersUI } from "../../../mapLayers";
+import { MapLayersUI } from "../../mapLayers";
 
 import type { ScreenViewport } from "@itwin/core-frontend";
-import type { StyleMapLayerSettings } from "../../Interfaces";
+import type { StyleMapLayerSettings } from "../Interfaces";
 interface MapLayerSettingsMenuProps {
   mapLayerSettings: StyleMapLayerSettings;
   onMenuItemSelection: (action: string, mapLayerSettings: StyleMapLayerSettings) => void;
